lab/vuln_apps/demo_text2sql_toolkit/pyproject.toml
[tool.poetry]
name = "demo-text2sql-toolkit"
version = "0.1.0"
description = ""
authors = ["Mithlesh Upadhyay <mithleshupadhyay515@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
gradio = "^5.16.0"
openai = "^1.63.0"
dtx-prompt-guard-client = ">=0.2.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

